refactor(core): rename platform/{bootstrap,config,registry} to core/{bootstrap,config,registry}
整合 platform/ 与 core/ 重叠职责：将 bootstrap、config、registry 三个子目录
从 platform/ 整体搬到 core/，与原有 core/{mapper,pagination} 合并为单一
聚合根。同步更新 servora 内部 3 处自引用 import。

包名不变（bootstrap / config / registry）。build 通过。
