+++ core/bootstrap/AGENTS.md
