+++ core/bootstrap/lifecycle_test.go
