+++ core/config/etcd_test.go
