+++ core/registry/etcd.go
