+++ core/registry/etcd_error_test.go
