+++ core/registry/etcd_test.go
