+++ core/registry/kubernetes.go
