+++ core/registry/kubernetes_test.go
