+++ core/registry/registry.go
